# telemetry/go/model_telemetry_value.go
/*
 * Телеметрия
 *
 * Описание синхронного API
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type TelemetryValue struct {

	Name string `json:"name,omitempty"`

	Value string `json:"value,omitempty"`
}